+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - info: -B instead of -b for private bridge information
 - copy-in: copy is executed in the jail environment, to avoid soft-link related issues in the destination path
 - copy-in: with running pots, a -F flag is needed to force the copy, an operation that is discouraged for security reasons
+- fetch base.txz: the base FreeBSD tarball used to be temporarily stored in /tmp. While a POT_CACHE folder is available, use that instead.
 
 ### Removed
 - create-dns: remove this already deprecated command, leaving the user to create a dns for the public bridge